requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.3
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dataclasses-json==0.5.7
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
exceptiongroup==1.1.1
filelock==3.10.7
firebase==4.0.1
firebase-admin==6.1.0
Flask==2.2.3
Flask-Cors==3.0.10
flask-swagger-ui==4.11.1
flask-swagger-generator==0.4
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gptcache==0.1.10
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.148
lazy-object-proxy==1.9.0
loguru==0.7.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
pinecone-client==2.2.1
platformdirs==3.2.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pylint==2.17.2
pyparsing==3.0.9
pystache==0.6.0
pytest==7.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
replicate==0.8.1
requests==2.28.2
rsa==4.9
six==1.16.0
slack==0.0.2
SQLAlchemy==1.4.47
tenacity==8.2.2
tiktoken==0.3.3
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.2.3
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.8.2
